+++ pkg/inline/inline.go
@@ -6,17 +6,17 @@ import (
 )
 
 type Weights struct {
-	Int          uint `xml:"int"`
-	Float        uint `xml:"float"`
-	Imag         uint `xml:"imag"`
-	Char         uint `xml:"char"`
-	String       uint `xml:"string"`
-	CompositeLit uint `xml:"composite"`
+	Int          float64 `xml:"int"`
+	Float        float64 `xml:"float"`
+	Imag         float64 `xml:"imag"`
+	Char         float64 `xml:"char"`
+	String       float64 `xml:"string"`
+	CompositeLit float64 `xml:"composite"`
 }
 
 type Metric struct {
 	Config Weights
-	Comp   uint
+	Comp   float64
 }
 
 func (m *Metric) ParseNode(n ast.Node) {
@@ -29,12 +29,12 @@ func (m *Metric) ParseNode(n ast.Node) {
 }
 
 func (m Metric) Finish() float64 {
-	return float64(m.Comp)
+	return m.Comp
 }
 
-func getBasicLitComp(literal *ast.BasicLit, config *Weights) uint {
-	var comp uint
-	len := uint(len(literal.Value))
+func getBasicLitComp(literal *ast.BasicLit, config *Weights) float64 {
+	var comp float64
+	len := float64(len(literal.Value))
 	switch literal.Kind {
 	case token.INT:
 		comp = config.Int
